feat(orchestrator): wire csi.Selector; delete internal/capi/csi
Bootstrap now iterates csi.Selector(cfg) to call EnsureSecret on each
registered driver instead of calling the Proxmox-only helpers in
internal/capi/csi directly. LoadVarsFromConfig and
ApplyConfigSecretToWorkload are moved into internal/csi/proxmoxcsi,
which owns all Proxmox CSI Secret logic. The now-empty internal/capi/csi
package is deleted.

Closes #118
